update wrong column calculation algorithm

+++ recyclercolumnadaptable/src/main/java/eu/codlab/recyclercolumnadaptable/item/AbstractItem.java
@@ -26,7 +26,7 @@ public abstract class AbstractItem {
         _position = position;
     }
 
-    public long getPosition() {
+    public int getPosition() {
         return _position;
     }
